+++ benchmarks/SkiaSharp.Benchmarks/Resources/LottieLogo1.json
