+++ benchmarks/SkiaSharp.Benchmarks/SkiaSharp.Benchmarks.csproj
@@ -34,4 +34,7 @@
   <ItemGroup>
     <None Remove="BenchmarkDotNet.Artifacts/**" />
   </ItemGroup>
+  <ItemGroup>
+    <Content Include="Resources\**" CopyToOutputDirectory="PreserveNewest" />
+  </ItemGroup>
 </Project>
