+++ benchmarks/SkiaSharp.Benchmarks/Utils/NonSeekableReadOnlyStream.cs
